refactor : 패키지 이름 변경 #4
패키지 이름 생성 규칙을 지키도록 변경
