+++ BE/src/main/java/com/codesquad/sidedish/domain/Category.java
@@ -1,4 +1,4 @@
-package com.codesquad.sidedish.entity;
+package com.codesquad.sidedish.domain;
 
 import org.springframework.data.annotation.Id;
