+++ BE/src/main/java/com/codesquad/sidedish/repository/CategoryRepository.java
@@ -1,6 +1,6 @@
 package com.codesquad.sidedish.repository;
 
-import com.codesquad.sidedish.entity.Category;
+import com.codesquad.sidedish.domain.Category;
 import org.springframework.data.repository.CrudRepository;
 
 public interface CategoryRepository extends CrudRepository<Category, Long> {
